+++ packages/koop-core/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [3.12.1] - 2019-06-04
 ### Changed
 * preferentially use the `config` argument for `this.config` during instantiation
 
@@ -846,6 +846,8 @@ Koop is now just a node module that exposes an express middleware app with hooks
   - to use Koop you must use it as middleware in an app that boots up an http server
 
 
+[3.12.1]: https://github.com/koopjs/koop-core/compare/v3.12.0...v3.12.1
+[3.12.0]: https://github.com/koopjs/koop-core/compare/v3.11.1...v3.12.0
 [3.11.1]: https://github.com/koopjs/koop-core/compare/v3.11.0...v3.11.1
 [3.11.0]: https://github.com/koopjs/koop-core/compare/v3.10.1...v3.11.0
 [3.10.1]: https://github.com/koopjs/koop-core/compare/v3.10.0...v3.10.1